+++ src/main/java/com/wztwzt/ae2_qof/network/WirelessHighlightPacket.java
@@ -3,19 +3,18 @@ package com.wztwzt.ae2_qof.network;
 import java.util.ArrayList;
 import java.util.List;
 
-import net.minecraft.client.Minecraft;
-
-import com.wztwzt.ae2_qof.client.ClientState;
+import com.wztwzt.ae2_qof.MyMod;
 
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
+import cpw.mods.fml.relauncher.Side;
 import io.netty.buffer.ByteBuf;
 
 public class WirelessHighlightPacket implements IMessage {
 
-    private List<int[]> positions;
-    private boolean enable;
+    public List<int[]> positions;
+    public boolean enable;
 
     public WirelessHighlightPacket() {
         this.positions = new ArrayList<int[]>();
@@ -66,16 +65,10 @@ public class WirelessHighlightPacket implements IMessage {
     public static class Handler implements IMessageHandler<WirelessHighlightPacket, IMessage> {
 
         @Override
-        public IMessage onMessage(final WirelessHighlightPacket message, MessageContext ctx) {
-            Minecraft.getMinecraft()
-                .func_152344_a(new Runnable() {
-
-                    @Override
-                    public void run() {
-                        ClientState.highlightPositions = message.positions;
-                        ClientState.highlightEnabled = message.enable;
-                    }
-                });
+        public IMessage onMessage(WirelessHighlightPacket message, MessageContext ctx) {
+            if (ctx.side == Side.CLIENT) {
+                MyMod.proxy.handleWirelessHighlight(message);
+            }
             return null;
         }
     }